# inflate: Inline symbol read