```
λ benchcmp old.txt new.txt
benchmark                       old ns/op     new ns/op     delta
BenchmarkGunzipCopy-32          22523140      19900003      -11.65%
BenchmarkGunzipNoWriteTo-32     22542379      19977600      -11.38%
BenchmarkGunzipStdlib-32        23444056      23435431      -0.04%

benchmark                       old MB/s     new MB/s     speedup
BenchmarkGunzipCopy-32          211.91       239.85       1.13x
BenchmarkGunzipNoWriteTo-32     211.73       238.91       1.13x
BenchmarkGunzipStdlib-32        203.59       203.66       1.00x
```